requirements/windows.txt
docutils
pygments
pypiwin32
kivy.deps.sdl2
kivy.deps.glew
kivy.deps.gstreamer
--extra-index-url https://kivy.org/downloads/packages/simple/

-r base.txt
